automation-system/llm2-solution/requirements.txt
flask==2.2.5
requests==2.31.0
python-dotenv==0.21.0
openai==0.28.0
